[back] 폴더 구조 변경, nodemon.json 추가, import 절대경로 설정
