+++ backend/nodemon.json
@@ -1,0 +1,6 @@
+{
+  "watch": ["src"],
+  "ext": "ts",
+  "ignore": ["node_modules", "src/**/*.spec.ts"],
+  "exec": "ts-node -r tsconfig-paths/register --files ./src/app.ts"
+}
