--- backend/src/adapter/infra/redis/usecase.ts
@@ -1,8 +1,0 @@
-import { createClient } from 'redis'
-import config from '../../config'
-
-export const redisClient = createClient({
-  url: `redis://${config.redis.host}:${config.redis.port}`,
-  legacyMode: true,
-})
-redisClient.connect().catch(console.error)
