--- backend/src/application/services/index.ts
@@ -1,1 +1,0 @@
-export { default as authService } from './auth.service'
