+++ backend/src/services/index.ts
@@ -1,0 +1,3 @@
+export { default as authService } from './auth.service'
+export { default as articleService } from './article.service'
+export { default as userService } from './user.service'
