+++ backend/tsconfig.json
@@ -14,8 +14,9 @@
     "esModuleInterop": true,
     "jsx": "react",
     "strictPropertyInitialization": false,
-    "emitDecoratorMetadata": true
+    "emitDecoratorMetadata": true,
+    "baseUrl": "src"
   },
-  "include": ["src/**/*"],
+  "include": ["src"],
   "exclude": ["node_modules", "**/*.spec.ts"]
 }
